Created tree.c, heap.c, added function templates
I gave you a struct to use in tree.c, but feel free to ignore it entirely. If you do, let me know what you change it to so I can change my functions accordingly. I set up some basic functions to check the tree and see if it is properly heapified.